Moving BrainVISET to a separate file and fixing imports.
